+++ script/enemy.ts
@@ -333,9 +333,9 @@ class Enemy extends Character {
       // @ts-ignore
       if (target) {
         if (generatePath(this.cords, target.cords, this.canFly, true) <= range) {
-          let encounter = player.entitiesEverEncountered?.enemies?.[this.indexInBaseArray.toString()];
+          let encounter = player.entitiesEverEncountered?.enemies?.[this.id];
           if(encounter < 1 || !encounter) {
-            player.entitiesEverEncountered.enemies[this.indexInBaseArray.toString()] = 1;
+            player.entitiesEverEncountered.enemies[this.id] = 1;
             displayText("New enemy encountered!");
             displayText(this.id + " added to codex.");
             spawnFloatingText(this.cords, "NEW ENEMY ENCOUNTER", "yellow", 22, 2000, 0);